refactor: extract dask-specific functions to dask.py
- Move dask_groupby_agg, dask_groupby_scan, _grouped_combine, and _unify_chunks to flox/dask.py
- Update imports in core.py and scan.py to use new dask module locations
- Keep reindex_intermediates in core.py as requested
- Maintain backward compatibility and library functionality
